split HashHelper from SecurityHelper
